Refactoring variants.models (#777)
Closes: #777
Related-Issue: #777
Projected-Results-Impact: none
