+++ variants/models/__init__.py
@@ -1,0 +1,18 @@
+"""The models and supporting code for the ``variants`` app.
+
+Split into smaller sub modules from which we import everything here.
+"""
+
+from variants.models.case import *  # noqa: F401 F403
+from variants.models.export import *  # noqa: F401 F403
+from variants.models.extsub import *  # noqa: F401 F403
+from variants.models.importer import *  # noqa: F401 F403
+from variants.models.kiosk import *  # noqa: F401 F403
+from variants.models.maintenance import *  # noqa: F401 F403
+from variants.models.projectroles import *  # noqa: F401 F403
+from variants.models.queries import *  # noqa: F401 F403
+from variants.models.scores import *  # noqa: F401 F403
+from variants.models.stats import *  # noqa: F401 F403
+from variants.models.summary import *  # noqa: F401 F403
+from variants.models.userannos import *  # noqa: F401 F403
+from variants.models.variants import *  # noqa: F401 F403
